+++ CHANGES.md
@@ -28,6 +28,9 @@
   flag as documented in the reference manual.
   ([#1641](https://github.com/GaloisInc/cryptol/issues/1641))
 
+* Add a REPL option `tcSmtFile` that allows writing typechecker-related SMT
+  solver interactions to a file.
+
 # 3.2.0 -- 2024-08-20
 
 ## Language changes